RD-1526 Extract deployment status icon components
Makes their reuse easier
